extract gob-web's database singleton
