+++ modules/gob-web-database/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@gob/web-database",
+  "version": "3.0.0-rc.2",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "@gob/treo-plugin-batch-get": "^3.0.0-rc.2",
+    "@gob/treo-plugin-query": "^3.0.0-rc.2",
+    "es6-promise": "^4.2.4",
+    "treo": "^0.6.0-rc2"
+  },
+  "devDependencies": {
+    "fake-indexeddb": "^2.0.4"
+  }
+}
